Conflicts:
	App/src/canvas.cpp

Former-commit-id: a5d5ae51f610ad3033314a4fb5ba6f3d05dfc42e
Former-commit-id: 98512e15db88aef45f3dc22a86c632ac5d8a997a

+++ App/src/canvas.cpp
@@ -31,7 +31,7 @@ void Canvas::initializeGL()
 	signalProcessor = new SignalProcessor(dataFile);
 
 	program = new OpenGLProgram(PROGRAM_OPTIONS->get("vert").as<string>().c_str(),
-	                            PROGRAM_OPTIONS->get("frag").as<string>().c_str());
+								PROGRAM_OPTIONS->get("frag").as<string>().c_str());
 
 	fun()->glGenVertexArrays(1, &vertexArray);
 	fun()->glBindVertexArray(vertexArray);
@@ -67,12 +67,12 @@ void Canvas::paintGL()
 	fun()->glUseProgram(program->getGLProgram());
 
 	GLuint location = fun()->glGetUniformLocation(program->getGLProgram(), "transformMatrix");
-	checkNotErrorCode(location, -1, "glGetUniformLocation() failed.");
+	checkNotErrorCode(location, static_cast<GLuint>(-1), "glGetUniformLocation() failed.");
 	fun()->glUniformMatrix4fv(location, 1, GL_FALSE, matrix.data());
 
 	// Create the data block range needed.
 	int firstIndex = static_cast<int>(floor(parent->getPosition()*ratio)),
-	    lastIndex = static_cast<int>(ceil((parent->getPosition()+width())*ratio));
+		lastIndex = static_cast<int>(ceil((parent->getPosition()+width())*ratio));
 
 	firstIndex /= signalProcessor->getBlockSize();
 	lastIndex /= signalProcessor->getBlockSize();
@@ -125,19 +125,19 @@ void Canvas::paintBlock(const SignalBlock& block)
 void Canvas::paintChannel(unsigned int channel, const SignalBlock& block)
 {
 	GLuint location = fun()->glGetUniformLocation(program->getGLProgram(), "y0");
-	checkNotErrorCode(location, -1, "glGetUniformLocation() failed.");
+	checkNotErrorCode(location, static_cast<GLuint>(-1), "glGetUniformLocation() failed.");
 	float y0 = (channel + 0.5f)*height()/block.getchannelCount();
 	fun()->glUniform1f(location, y0);
 
 	location = fun()->glGetUniformLocation(program->getGLProgram(), "yScale");
-	checkNotErrorCode(location, -1, "glGetUniformLocation() failed.");
+	checkNotErrorCode(location, static_cast<GLuint>(-1), "glGetUniformLocation() failed.");
 	fun()->glUniform1f(location, -0.000008f*height());
 
 	GLint size = block.getLastSample() - block.getFirstSample() + 1;
 	GLint first = channel*size;
 
 	location = fun()->glGetUniformLocation(program->getGLProgram(), "bufferOffset");
-	checkNotErrorCode(location, -1, "glGetUniformLocation() failed.");
+	checkNotErrorCode(location,static_cast<GLuint>(-1), "glGetUniformLocation() failed.");
 	fun()->glUniform1i(location, block.getFirstSample() - first);
 
 	fun()->glDrawArrays(GL_LINE_STRIP, first , size);
